% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParCC-package.R
\docType{package}
\name{ParCC-package}
\alias{ParCC}
\alias{ParCC-package}
\title{ParCC: Parameter Converter and Calculator for HTA}
\description{
A decision support tool for Health Technology Assessment (HTA). Bridges the gap between clinical evidence and economic models by automating complex parameter estimations including rate-probability conversions, hazard ratio-based probability conversion, survival extrapolation, PSA distribution fitting, and value-based pricing analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/drpakhare/ParCC}
  \item Report bugs at \url{https://github.com/drpakhare/ParCC/issues}
}

}
\author{
\strong{Maintainer}: RRC-HTA Team \email{hta.aiimsbhopal@gmail.com}

}
\keyword{internal}
